+++ submission/src/epilot/sdk.py
@@ -4,6 +4,7 @@ import requests as requests_http
 from .sdkconfiguration import SDKConfiguration
 from .submissions import Submissions
 from epilot import utils
+from typing import Dict
 
 class Epilot:
     r"""Submission API: Use this API to handle submissions entities from external sources e.g. journeys and frontends"""
@@ -15,7 +16,7 @@ class Epilot:
     def __init__(self,
                  server_idx: int = None,
                  server_url: str = None,
-                 url_params: dict[str, str] = None,
+                 url_params: Dict[str, str] = None,
                  client: requests_http.Session = None,
                  retry_config: utils.RetryConfig = None
                  ) -> None:
@@ -26,7 +27,7 @@ class Epilot:
         :param server_url: The server URL to use for all operations
         :type server_url: str
         :param url_params: Parameters to optionally template the server URL with
-        :type url_params: dict[str, str]
+        :type url_params: Dict[str, str]
         :param client: The requests.Session HTTP client to use for all operations
         :type client: requests_http.Session
         :param retry_config: The utils.RetryConfig to use globally